各画像付きカードをImageCardとして別コンポーネントに分ける
