+++ pages/index.tsx
@@ -21,11 +21,7 @@ import Link from '@mui/material/Link';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 
 import Copyright from '../components/Copyright';
-
-function getRandom(min: number=0, max: number=1000): number {
-    var random = Math.floor( Math.random() * (max + 1 - min) ) + min;
-    return random;
-}
+import ImageCard from '../components/ImageCard';
 
 const inter = Inter({ subsets: ['latin'] })
 const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9];
@@ -96,36 +92,9 @@ export default function Album() {
           {/* End hero unit */}
           <Grid container spacing={4}>
             {cards.map((card) => {
-              const rand = getRandom()
-              console.log(rand)
               return (
               <Grid item key={card} xs={12} sm={6} md={4}>
-                <Card
-                  sx={{ height: '100%', display: 'flex', flexDirection: 'column' }}
-                >
-                  <CardMedia
-                    component="img"
-                    sx={{
-                      // 16:9
-                      pt: '56.25%',
-                    }}
-                    image={"https://source.unsplash.com/random?" + getRandom()}
-                    alt="random"
-                  />
-                  <CardContent sx={{ flexGrow: 1 }}>
-                    <Typography gutterBottom variant="h5" component="h2">
-                      Heading
-                    </Typography>
-                    <Typography>
-                      This is a media card. You can use this section to describe the
-                      content.
-                    </Typography>
-                  </CardContent>
-                  <CardActions>
-                    <Button size="small">View</Button>
-                    <Button size="small">Edit</Button>
-                  </CardActions>
-                </Card>
+                <ImageCard />
               </Grid>
               )
             })}
